cursor/requirements.txt
Flask
pillow
python-barcode
openpyxl
reportlab